queue: split it in a few modules
The queue module is a bit crowded. It is easier to reason and extend by
slitting things a little bit. Extract out error, and move queue,
descriptor and descriptor_chain in a submodule.

A few fields and constants have been made pub(crate).
